refactor(assignments-service): Move constants into a shared file
